+++ Sources/KeyboardKit/Extensions/Font+Weight.swift
@@ -3,6 +3,7 @@
 //  KeyboardKit
 //
 //  Created by Brennan Drew on 2021-01-16.
+//  Copyright © 2021 Daniel Saidi. All rights reserved.
 //
 
 import UIKit